Refactor ShowNetworkLogo.py -> network_logo.py
