+++ medusa/media/network_logo.py
