+++ medusa/server/api/v1/core.py
@@ -48,7 +48,7 @@ from ....helper.quality import get_quality_string
 from ....logger import filter_logline, read_loglines
 from ....media.ShowBanner import ShowBanner
 from ....media.ShowFanArt import ShowFanArt
-from ....media.ShowNetworkLogo import ShowNetworkLogo
+from ....media.network_logo import ShowNetworkLogo
 from ....media.poster import ShowPoster
 from ....search.queue import BacklogQueueItem, ForcedSearchQueueItem
 from ....show.coming_episodes import ComingEpisodes
