+++ medusa/server/web/core/base.py
@@ -30,7 +30,7 @@ from .... import (
 )
 from ....media.ShowBanner import ShowBanner
 from ....media.ShowFanArt import ShowFanArt
-from ....media.ShowNetworkLogo import ShowNetworkLogo
+from ....media.network_logo import ShowNetworkLogo
 from ....media.poster import ShowPoster
 from ....show.coming_episodes import ComingEpisodes
