+++ pytest.ini
@@ -200,7 +200,7 @@ flake8-ignore =
     medusa/media/__init__.py D104
     medusa/media/ShowBanner.py D100 D102 D200 D400
     medusa/media/ShowFanArt.py D100 D102 D200 D400
-    medusa/media/ShowNetworkLogo.py D100 D102 D200 D400
+    medusa/media/network_logo.py D100 D102 D200 D400
     medusa/media/poster.py D100 D102 D200 D400
     medusa/providers/exceptions.py D200 D400
     medusa/providers/generic_provider.py D100 D101 D102 D202 D205 D400 D401 E501 F841
